dev_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
# The SDK in this mono-repo should be installed in editable mode:
# pip install -e ../sdk/python
# For on-chain payments the SDK requires web3/eth-account etc. See sdk/python/pyproject.toml
